PXMB-000 move Jwt classes to a jwt folder
